Rename routers modules
