+++ app/routers/notifications_router.py
@@ -8,7 +8,7 @@ from sqlalchemy.orm import Session
 from app.core.database import get_db
 from app.models.notification import Notification
 from app.models.user import User
-from app.routers.users import get_current_user
+from app.routers.users_router import get_current_user
 from app.schemas.notification import NotificationOut  # <-- import BEFORE usage
 
 router = APIRouter(prefix="/notifications", tags=["notifications"])
