+++ app/routers/tasks_router.py
@@ -7,7 +7,7 @@ from app.core.database import get_db
 from app.models.task import Task, Priority, Status
 from app.models.user import User, Role
 from app.schemas.task import TaskCreate, TaskUpdate, TaskOut
-from app.routers.users import get_current_user
+from app.routers.users_router import get_current_user
 
 router = APIRouter(prefix="/tasks", tags=["tasks"])
